+++ .idea/modules/hazthenewestapp.test.iml
@@ -2,13 +2,13 @@
 <module version="4">
   <component name="FacetManager">
     <facet type="kotlin-language" name="Kotlin">
-      <configuration version="4" platform="JavaScript " allPlatforms="JS []" useProjectSettings="false" isTestModule="true" externalProjectId="hazthenewestapp" pureKotlinSourceFolders="C:\Users\safar\Documents\GitHub\hazthenewestapp\src\test\kotlin;C:\Users\safar\Documents\GitHub\hazthenewestapp\src\main\kotlin;C:\Users\safar\Documents\GitHub\hazthenewestapp\build\externals\hazthenewestapp\src;C:\Users\safar\Documents\GitHub\hazthenewestapp\src\main\kotlin;C:\Users\safar\Documents\GitHub\hazthenewestapp\build\externals\hazthenewestapp\src;C:\Users\safar\Documents\GitHub\hazthenewestapp\src\test\kotlin">
+      <configuration version="4" platform="JavaScript " allPlatforms="JS []" useProjectSettings="false" isTestModule="true" externalProjectId="hazthenewestapp">
         <sourceSets>
           <sourceSet>hazthenewestapp.main</sourceSet>
         </sourceSets>
         <newMppModelJpsModuleKind>COMPILATION_AND_SOURCE_SET_HOLDER</newMppModelJpsModuleKind>
         <externalSystemTestTasks>
-          <externalSystemTestTask>browserTest|hazthenewestapp:test|js</externalSystemTestTask>
+          <externalSystemTestTask>browserTest|:|js</externalSystemTestTask>
         </externalSystemTestTasks>
         <testOutputPath>$MODULE_DIR$/../../build/classes/kotlin/test</testOutputPath>
         <compilerSettings>
@@ -31,7 +31,6 @@
           <option name="pluginClasspaths">
             <array />
           </option>
-          <option name="multiPlatform" value="true" />
           <option name="errors">
             <ArgumentParseErrors />
           </option>
